Alteracoes nas q1, q3 e q4

+++ Eron_Morais/Exercicios_python/02EstruturaDeDecisao/q01.py
@@ -1,7 +1,7 @@
 #Faça um Programa que peça dois números e imprima o maior deles
 
-n1 = int(input('Digite um numero:'))
-n2 = int(input('Digite outro numero:'))
+n1 = float(input('Digite um numero:'))
+n2 = float(input('Digite outro numero:'))
 
 if n1 > n2:
     print(n1)
